Fix plugin load errors caused by yapsy 1.12
Yapsy’s latest version uses its new plugin loading system, which I mentioned in a comment in main.py. This plugin loading system behaves differently from the way I was expecting it to work, and so the “automatic upgrade” wasn’t so automatic, or much of an upgrade.

NOTE: Fixing this required that I rename all of the plugin info files and python files to match the name of the plugin’s main class. This also changes the plugin name in the configuration file.